+++ src/babel.ts
@@ -1,9 +1,6 @@
-import { pathToFileURL } from 'url'
-import { smart } from '@babel/template'
-import type { NodePath, PluginObj } from '@babel/core'
-import type { Statement, MemberExpression } from '@babel/types'
 import { transformSync, TransformOptions as BabelTransformOptions } from '@babel/core'
 import { TransformOptions, TRANSFORM_RESULT } from './types'
+import { TransformImportMetaPlugin } from './plugins/babel-plugin-transform-import-meta'
 
 export default function transform (opts: TransformOptions): TRANSFORM_RESULT {
   const _opts: BabelTransformOptions = {
@@ -52,46 +49,6 @@ export default function transform (opts: TransformOptions): TRANSFORM_RESULT {
         code: err.code?.replace('BABEL_', '').replace('PARSE_ERROR', 'ParseError'),
         message: err.message?.replace('/: ', '').replace(/\(.+\)\s*$/, '')
       })
-    }
-  }
-}
-
-// https://github.com/javiertury/babel-plugin-transform-import-meta/blob/master/src/index.ts
-const TransformImportMetaPlugin = (opts: { filename?: string }): PluginObj => {
-  return {
-    name: 'transform-import-meta',
-    visitor: {
-      Program (path) {
-        const metas: Array<NodePath<MemberExpression>> = []
-
-        path.traverse({
-          MemberExpression (memberExpPath) {
-            const { node } = memberExpPath
-
-            if (
-              node.object.type === 'MetaProperty' &&
-              node.object.meta.name === 'import' &&
-              node.object.property.name === 'meta' &&
-              node.property.type === 'Identifier' &&
-              node.property.name === 'url'
-            ) {
-              metas.push(memberExpPath)
-            }
-          }
-        })
-
-        if (metas.length === 0) {
-          return
-        }
-
-        for (const meta of metas) {
-          meta.replaceWith(smart.ast`${
-            opts.filename
-              ? JSON.stringify(pathToFileURL(opts.filename))
-              : "require('url').pathToFileURL(__filename).toString()"
-          }` as Statement)
-        }
-      }
     }
   }
 }
